refactor so gun only does what it needs
