+++ src-ui/deps.js
@@ -1,14 +1,13 @@
-import Gun from 'gun';
+import { gun as database } from './src/system/database.js'
 import Color from "colorjs.io";
 import { Midi } from '@tonejs/midi'
 import * as Tone from 'tone'
 
-import * as focusTrap from 'focus-trap';
-import module from './src/system/module.js'
-import devices from './src/system/devices.js'
+import './src/system/devices.js'
 
-const gun = Gun(['https://gun-manhattan.herokuapp.com/gun'])
-const fgun = foot(gun, window.location.pathname)
+import * as focusTrap from 'focus-trap';
+
+router(database, window.location.pathname)
 
 export {
   Color,
@@ -16,17 +15,13 @@ export {
   Tone,
   focusTrap,
   randomString,
-  module,
-	foot,
-  fgun,
-  devices
 }
 
 const randomString = (length) =>
   [ ...Array(length) ].map(() => (~~(Math.random() * 36)).toString(36)).join('');
 
-function foot(gun, path) {
-	const fgun = gun.get('grapevine').get(path)
+function router(database, path) {
+	const file = database.get('files').get(path)
 
 	const router = {
 		'/edit/': editHandler,
@@ -45,8 +40,6 @@ function foot(gun, path) {
 		standardHandler(path)
 	}
 
-	return fgun
-
 	function editHandler() {
 		render('edit', {class: 'world'}, "edit")
 	}
@@ -56,7 +49,9 @@ function foot(gun, path) {
 	}
 
 	function standardHandler() {
-		render('main-stickies')
+    if(window.top == self.self) {
+      render('main-stickies')
+    }
 	}
 
 	function render(element, properties={}, innards='') {
