+++ src-ui/src/synth-module.js
@@ -1,5 +1,6 @@
 import module from './system/module.js'
-import { Color, Tone, Midi, devices } from '../deps.js'
+import devices from './system/devices.js'
+import { Color, Tone, Midi } from '../deps.js'
 import $guitar from "./guitar.js"
 
 const start = new Date()
