+++ src-ui/src/system/database.js
@@ -1,0 +1,3 @@
+import Gun from 'gun';
+
+export const gun = Gun(['https://gun-manhattan.herokuapp.com/gun'])
